# Minimally invasive changes to improve search performance.