@chmorgan 2024-04-09
@whmountains I'll mention that my approach mostly obviates the part search guide elements other than the 'keywords' list.

I had considered this approach and you are correct that a massive amount of performance can be gained. The downside though is that you end up having to do a lot of the advanced query construction manually. The database is much smaller though, the fts5 database is a hefty 4GB. The zip size is only ~700MB though (so 3x the size).

My workflow of finding parts is to combine the jlcpcb plugin with the website and plug in freeform entries to the keywords list.

I've updated #348 against latest main and pushed out a new db to my GitHub hosted site if you want to give it a spin.