@whmountains 2024-04-09
I love the idea of having a single search box.  Thanks for releasing the update, I will test it out and report back.